+++ oldPackage.json
@@ -1,0 +1,33 @@
+{
+  "name": "with-typescript",
+  "version": "1.0.0",
+  "scripts": {
+  },
+  "dependencies": {
+    "@apollo/react-hooks": "^3.1.5",
+    "@material-ui/core": "^4.9.14",
+    "@material-ui/icons": "^4.9.1",
+    "@material-ui/styles": "^4.9.14",
+    "@types/react-redux": "^7.1.8",
+    "apollo-cache-inmemory": "^1.6.6",
+    "apollo-client": "^2.6.10",
+    "apollo-link-http": "^1.5.17",
+    "apollo-server": "^2.13.1",
+    "apollo-server-lambda": "^2.13.1",
+    "graphql": "^15.0.0",
+    "isomorphic-unfetch": "3.0.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-redux": "^7.2.0",
+    "react-spring": "^8.0.27",
+    "redux": "^4.0.5",
+    "redux-devtools-extension": "^2.13.8"
+  },
+  "devDependencies": {
+    "@types/node": "^14.0.1",
+    "@types/react": "^16.9.35",
+    "@types/react-dom": "^16.9.8",
+    "typescript": "3.9.2"
+  },
+  "license": "ISC"
+}
