--- pages/About.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import Layout from '../components/Layout';
-
-export default function About() {
-  return (
-    <Layout title="Portara | About">
-      This is just some text on our About page
-    </Layout>
-  );
-}
