--- pages/Documentation.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import Layout from '../components/Layout';
-
-export default function Documentation() {
-  return (
-    <Layout title="Portara | Documentation">
-      This is just some text on our Documentation page
-    </Layout>
-  );
-}
