--- pages/api/graphql/resolvers.ts
@@ -1,7 +1,0 @@
-
-
-export const resolvers = {
-  Query: {
-    // insert queries
-  },
-};
