--- pages/api/graphql/typeDefs.ts
@@ -1,10 +1,0 @@
-
-
-const { gql } = require('apollo-server-lambda');
-
-export const typeDefs = gql`
-  type Query {
-    # insert queries
-  },
-`;
-
