--- serverless.yml
@@ -1,22 +1,0 @@
-# serverless.yml
-
-#org: portara
-#app: portara-website
-service: apollo-lambda
-provider:
-  name: aws
-  runtime: nodejs12.x
-  region: us-east-2
-functions:
-  graphql:
-    # this is formatted as <FILENAME>.<HANDLER>
-    handler: server.graphqlHandler
-    events:
-    - http:
-        path: graphql
-        method: post
-        cors: true
-    - http:
-        path: graphql
-        method: get
-        cors: true
